azure-functions/requirements.txt
azure-functions
requests==2.22.0
azure-storage-blob==12.3.0
azure-identity==1.7.1